requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
bcrypt==4.0.1
boto3==1.26.22
botocore==1.29.22
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.15
django-filter==22.1
django-storages==1.13.1
google-auth==2.11.1
google-auth-oauthlib==0.5.3
gspread==5.5.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.4
oauthlib==3.2.1
packaging==21.3
paramiko==2.11.0
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
transliterate==1.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.10.0
zope.interface==5.5.0
